refactor(ecsm): move inlined tests + reference arithmetic into src/tests/
The three source files mixed production code with test-only code. Relocate it all
into a dedicated src/tests/ tree (matching the stark/prover convention:
`#[cfg(test)] mod tests;` + tests/mod.rs), leaving lib.rs/curve.rs/witness.rs as
pure production:

- tests/lib_tests.rs     <- lib.rs's inlined `mod tests`
- tests/curve_tests.rs   <- curve.rs's `mod parity_tests`
- tests/witness_tests.rs <- witness.rs's inlined `mod tests`
- tests/reference.rs     <- curve.rs's #[cfg(test)] reference impl
                            (point_double / point_add / step_lambda /
                            replay_double_and_add_reference)
- tests/reference_field.rs <- the whole #[cfg(test)] field.rs (BigUint Fp)
- field.rs deleted

Test helpers' hex-parse .unwrap() -> .expect(...) for clearer panics. Fixed a
now-dangling intra-doc link in replay_double_and_add. Behavior-preserving:
ecsm lib 15/15, clippy clean, production builds with no test deps.
